requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.335
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
jinja2==3.1.2

# Frontend
streamlit==1.29.0

# Groq integration
# Note: Groq uses the OpenAI SDK, so no additional packages are needed

# LocalAI support
requests>=2.28.0
